Grab-bag of MSVC conversion warning fixes
Mostly double-to-float and int-to-float “possible loss of precision” —
fixed by changing constants to be float or inserting explicit casts.
None look at risk of meaningful failure, except for the
float_16_t_aottest, which was actually just wrong.

+++ test/generator/mandelbrot_aottest.cpp
@@ -10,7 +10,7 @@ using namespace Halide;
 int main(int argc, char **argv) {
     Buffer<int> output(100, 30);
     const char *code = " .:-~*={}&%#@";
-    const int iters = strlen(code) - 1;
+    const int iters = (int) strlen(code) - 1;
 
     // Compute 100 different julia sets
     for (float t = 0; t < 100; t++) {
